requirements.txt
flask
flask-sse
redis
kafka-python